fix(panama): fix arithmetic cast and instanceof issues
Fix several SpotBugs issues related to incorrect type casting:

1. ICAST_IDIV_CAST_TO_DOUBLE: Cast to double before division to prevent
   precision loss in PanamaBatchProcessor and PanamaConcurrencyManager
2. ICAST_INTEGER_MULTIPLY_CAST_TO_LONG: Cast to long before multiplication
   to prevent overflow in PanamaBatchProcessor
3. BAD_HEXA_CONVERSION: Use String.format("%02x") for proper hex conversion
   with leading zeros in CompilationCache
4. BC_VACUOUS_INSTANCEOF: Remove unnecessary instanceof check in
   PanamaWasmThread since type is known

These fixes prevent potential precision loss, overflow, hash collisions,
and unnecessary runtime checks.

+++ wasmtime4j-panama/src/main/java/ai/tegmentum/wasmtime4j/panama/util/PanamaBatchProcessor.java
@@ -251,7 +251,7 @@ public final class PanamaBatchProcessor {
 
     final long startTime = PanamaPerformanceMonitor.startOperation("panama_native_batch_process");
     try (Arena tempArena = Arena.ofConfined()) {
-      PanamaPerformanceMonitor.recordArenaAllocation(tempArena, parameters.length * 64);
+      PanamaPerformanceMonitor.recordArenaAllocation(tempArena, (long) parameters.length * 64);
 
       LOGGER.fine(
           () -> String.format("Processing Panama native batch: %d parameters", parameters.length));
@@ -384,7 +384,8 @@ public final class PanamaBatchProcessor {
     final long totalTimeNs = totalBatchTimeNs.get();
 
     final double avgBatchSize = batches > 0 ? (double) operations / batches : 0.0;
-    final double avgBatchTimeMs = batches > 0 ? (totalTimeNs / batches) / 1_000_000.0 : 0.0;
+    final double avgBatchTimeMs =
+        batches > 0 ? ((double) totalTimeNs / batches) / 1_000_000.0 : 0.0;
     final double avgOpTimeNs = operations > 0 ? (double) totalTimeNs / operations : 0.0;
 
     return String.format(
